Revert "Split side-comments cache into its own collection"
