+++ packages/lesswrong/server/manualMigrations/index.ts
@@ -118,5 +118,4 @@ import './2023-01-13-createViewUpdater'
 import './2023-01-18-user-mentions'
 import './2023-02-10-fillCreatedAtForDebouncerEvents'
 import './2023-03-16-createUserActivities'
-import './2023-04-10-importACXmeetups.ts'
-import './2023-06-04-populateSideCommentsCache'
+import './2023-04-10-importACXmeetups'
